requirements.txt
google-genai>=1.0.0
groq>=0.11.0
cohere>=5.11.0
SpeechRecognition>=3.10.0
gTTS>=2.5.0
pygame>=2.6.0
python-dotenv>=1.0.0
pyaudio>=0.2.14
